requirements.txt
XlsxWriter==1.1.2
requests==2.19.1
xlrd==1.2.0
xlwt==1.3.0
configparser==3.5.0
psycopg2==2.7.6.1
requests-toolbelt==0.8.0
rsa==4.0
PyExecJS==1.5.1
paramiko==2.4.2
WMI==1.4.9
pywin32==224
selenium==3.13.0